@@ -1,7 +1,7 @@
 from pathlib import Path
 from typing import Optional
-from instaui.zero.options import CdnResourceOption
-from instaui_tdesign import consts
+from instaui.cdn.options import CdnResourceOption
+from instaui_tdesign import resources
 
 
 def override(
@@ -18,13 +18,13 @@ def override(
 
     css_links: dict[Path, str] = {}
     if tdesign_vue_next_css:
-        css_links[consts.tdesign_css] = tdesign_vue_next_css
+        css_links[resources.tdesign_css] = tdesign_vue_next_css
 
     return CdnResourceOption(import_maps=import_maps, css_links=css_links)
 
 
 def default_override() -> CdnResourceOption:
     return override(
-        tdesign_vue_next_js=consts.TDESIGN_VUE_ESM_JS_CDN,
-        tdesign_vue_next_css=consts.TDESIGN_VUE_CSS_CDN,
+        tdesign_vue_next_js=resources.TDESIGN_VUE_ESM_JS_CDN,
+        tdesign_vue_next_css=resources.TDESIGN_VUE_CSS_CDN,
     )
