@@ -1,6 +1,6 @@
 from __future__ import annotations
 import typing
-from instaui.event.event_mixin import EventMixin
+from instaui.internal.ui.event import EventMixin
 from typing_extensions import TypedDict, Unpack
 from ._utils import handle_props, handle_event_from_props
 from ._base_element import BaseElement
