@@ -4,7 +4,7 @@ import typing
 from instaui_tdesign.components._icon_param_utils import make_icon_for_str
 
 from ._base_element import BaseElement
-from instaui.event.event_mixin import EventMixin
+from instaui.internal.ui.event import EventMixin
 from typing_extensions import TypedDict, Unpack
 from ._utils import handle_props, handle_event_from_props
 
